+++ runner.js
@@ -1,0 +1,5 @@
+let backpat = require('./index');
+
+// Run this script to test Backpat's output
+
+backpat((result) => console.log(result));