+++ test.js
@@ -1,3 +1,0 @@
-let deps = require('./index');
-
-deps((result) => console.log(result));